+++ .idea/workspace.xml
@@ -2,14 +2,19 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="fcb9d064-beec-4886-a293-23ab1b843534" name="Default Changelist" comment="">
-      <change afterPath="$PROJECT_DIR$/routes/setting.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/controllers/NewslettersController.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/controllers/SettingController.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/controllers/TagsController.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/controllers/UsersController.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/routes/tags.js" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/routes/users.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/app.js" beforeDir="false" afterPath="$PROJECT_DIR$/app.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/models/Newsletters.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/Newsletters.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/models/Setting.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/Setting.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/models/Tags.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/Tags.js" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/models/Users.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/Users.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/models/Newsletters.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/NewslettersModel.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/models/Setting.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/SettingModel.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/models/Tags.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/TagsModel.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/models/Users.js" beforeDir="false" afterPath="$PROJECT_DIR$/models/UsersModel.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/routes/newsletters.js" beforeDir="false" afterPath="$PROJECT_DIR$/routes/newsletters.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/routes/setting.js" beforeDir="false" afterPath="$PROJECT_DIR$/routes/setting.js" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/sequelize.js" beforeDir="false" afterPath="$PROJECT_DIR$/sequelize.js" afterDir="false" />
     </list>
     <option name="EXCLUDED_CONVERTED_TO_IGNORED" value="true" />
